Adding a bunch of small projects to github.
Adding some projects to github in their own little spot.

Regexbot.go is an IRC bot, written in golang.

shorts is a collection of programs written to be quite terse, to put it
lightly.

pathfinder.go is a means of finding the maximum path distance in a graph
where you can only traverse any given edge once.

sorts.go is simply a collection of various sorting algorithms.